% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxBuild.R
\name{boxBuild}
\alias{boxBuild}
\title{boxBuild()}
\usage{
boxBuild(data, xVar, yVar)
}
\arguments{
\item{data}{a data set to use}

\item{xVar}{x variable for plot, numeric variable}

\item{yVar}{y variable for plot, numeric variable}
}
\value{
A graph showing your results
}
\description{
The boxBuild() function allows you to easily create a box plot
}
\examples{
boxBuild(data = iris, xVar = iris$Sepal.Length, yVar = iris$Sepal.Width)
}
