% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\name{dimension}
\alias{dimension}
\title{Get the dimensions of a data set}
\usage{
dimension(data)
}
\arguments{
\item{data}{a data set to use}
}
\value{
print statements about your data
}
\description{
Get the dimensions of a data set
}
\examples{
dimension(iris)
}
