% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\name{dimension}
\alias{dimension}
\title{dimension()}
\usage{
dimension(data)
}
\arguments{
\item{data}{a data set to use}
}
\value{
print statements about your data
}
\description{
dimension() is a function that allows you to see the dimensions of your data
}
\examples{
dimension(iris)
}
