% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphBuild.R
\name{graphBuild}
\alias{graphBuild}
\title{graphBuild()}
\usage{
graphBuild(data, xVar, yVar)
}
\arguments{
\item{data}{a data set to use}

\item{xVar}{x variable for plot, numeric variable}

\item{yVar}{y variable for plot, numeric variable}
}
\value{
A graph showing your results
}
\description{
The graphBuild() function allows you to easily create a scatter plot
}
\examples{
graphBuild(data = iris, xVar = iris$Sepal.Length, yVar = iris$Sepal.Width)
}
